% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionstability.R
\name{geq}
\alias{geq}
\title{geq}
\usage{
geq(P, Q)
}
\arguments{
\item{P}{A partition of length n}

\item{Q}{Another partition of length n#'}
}
\description{
Implementation of Algorithm 4, which tests \eqn{P \geq Q}
}
\examples{
P = c(1, 1, 1, 2, 2, 2)
P_prime = c(2, 2, 2, 1, 1, 1)
Q = c(4, 4, 1, 3, 3, 2)
R = c(4, 4, 4, 4, 2, 2)
isTRUE(geq(P, Q))
isTRUE(geq(P, P_prime))
isTRUE(geq(P, P_prime) == geq(P_prime, P))
!isTRUE(geq(P, R))
!isTRUE(geq(R, P))

}
