% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionstability.R
\docType{package}
\name{partitionstability-package}
\alias{partitionstability}
\alias{partitionstability-package}
\title{partitionstability: Provides Implementations of Algorithms Presented in the Article of Ball and
Geyer-Schulz (2018)}
\description{
Exemplary implementations of Algorithms 1 to 4 of Ball, F and Geyer-Schulz A
(2018) <doi:TBA> that can be used to test partition stability given a set of
generators for a permutation group that acts on the partitioned set.
The implementations strongly relate to the pseudocode, which means there certainly
exist ways to implement the algorithms more efficiently.
However, the main focus lies on supplying supplemental material for the cited article.
}
\details{
Algorithms 1 (\link{testStability}) and 4 (\link{geq}) are nearly one-to-one 
implementations of the pseudocode. One exception is that we removed the need of a
hash map in \link{geq} by simply using a vector of the same length as the partition
to build our map of cluster ids. However, this implies that cluster ids must be the 
same as the node ids.

Algorithms 2 and 3 are merged into a single function (\link{computeOrbitPartition}) 
as R does not easily support call be reference.

For details of the algorithms and the background we refer to 
Ball, Fabian and Geyer-Schulz, Andreas (2018), 
``Symmetry-based Graph Clustering Partition Stability'', 
Archives of Datascience Series A <vol tba>(<nr tba>), DOI: <doi tba>
}
\author{
\strong{Maintainer}: Fabian Ball \email{fabian.ball@kit.edu} [copyright holder, contributor]

}
