% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionstability.R
\name{testStability}
\alias{testStability}
\title{testStability}
\usage{
testStability(P, S)
}
\arguments{
\item{P}{A partition}

\item{S}{Set of generators for a group \eqn{\langle S\rangle} that acts on \eqn{P}}
}
\description{
Implementation of Algorithm 1, which checks if a partition is stabl udner the given
symmetry implied by the generators of a permutation group \eqn{S}.
}
\examples{
p1 <- c(2L, 1L, 3L, 4L, 6L, 5L, 7L, 8L, 9L, 10L)
p2 <- c(1L, 2L, 3L, 4L, 5L, 6L, 7L, 8L, 10L, 9L)
p3 <- c(1L, 9L, 3L, 4L, 5L, 6L, 7L, 8L, 2L, 10L)
S <- list(p1, p2, p3)
P <- rep(1L, 10)
isTRUE(testStability(P, S))
Q <- c(2L, 2L, 1L, 1L, 1L, 1L, 1L, 1L, 3L, 3L)
!isTRUE(testStability(Q, S))

}
